Update patch set 1
Patch Set 1: Code-Review-1

I prefer you to move this task into the new play within the tasks directory.
Call it "clear_env.yml".
Move there the delete instance task.
Both of the tasks could be triggered by true/false condition.

The boot vm play could be used for booting vm not as a part of perf play, so it could be that I would like to keep the already created key and use it for other vm's to boot.

In the future tasks, I would like to add the ability to provide your own private key, and change the keypair name to variable.

Patch-set: 1
Label: Code-Review=-1