Update patch set 2
Patch Set 2:

Basically, pip already installed when python 2>= 2.7 or python 3 is installed.
But in case I would like to ensure it is installed and I want to use yum for the installation, python-pip package will be available to me only if an epel repository is installed.

Easy_install comes as part of python-setuptools which are available in any repository.
In such way, I can ensure pip is installed on my system without the need to install additional repositories.

Patch-set: 2