Add NFV lab switches configuration plays
NFV lab has two switches to manage.
1 - Provision switch
2 - Performance switch

The playbooks uses the defined switches port configuration to create a
required vlan and attach the required servers to that vlan.
Multiple servers could be supplied for the play.
Example within the docs.

The playbook will be used as part of the automation during the CI jobs.

New roles:
- Junos configuration role.
- Cisco configuration role.

Change-Id: I83813c72e81a45fe3cd8ef4384dcf0916814e317