Add NFV lab switches configuration play
NFV lab has two switches to manage.
1 - Provision switch (Cisco)
2 - Performance switch (Juniper)

The playbook uses the defined switches port configuration to create a
required vlan and attach the required servers to that vlan.
Multiple servers could be provided for the play.
Multiple switches could be provided for the play.
Examples within the docs.

The playbook will be used as part of the automation during the CI jobs.

New role:
- nfv_lab_switches

The role support Cisco and Juniper switches.

Change-Id: I83813c72e81a45fe3cd8ef4384dcf0916814e317