Change Tempest executor from os-testr to stestr
The stestr [1] is a fork of the testrepository that concentrates on being
a dedicated test runner for python projects.
The ostestr [2] command provides a wrapper around the testr command
included in the testrepository package.

In the latest ostestr release (1.0.0), one of the changes switch to
stestr under the covers [3].
As specified within the commit message of the change [3]:
  | This commit switches the ostestr command to use stestr
  | under the covers. This simplifies the majority of the logic,
  | eventually ostestr will be deprecated as a result of this
  | migration. (since almost all the logic here is contained
  | in stestr already)

Removing the wrapper layer of the ostestr and starting to use the
stestr.
Tested on the OSP 10 and OSP 13.

[1] - http://stestr.readthedocs.io/en/latest/README.html#overview
[2] - https://docs.openstack.org/os-testr/latest/user/ostestr.html
[3] - https://review.openstack.org/#/c/488441/

Change-Id: Ia6800d8c7fab4c48e128aa9860ee14663e128445