Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> (1 comment)

From my personal experience, Dell's modules throw an exception when you attempt to run them remotely, setting the connection to local doesn't fix it and the task must be delegated to localhost in order to work.

The creation of in-memory inventory is done for convenience, the play aims to be user friendly and using inventory provides the best feedback for users.

Patch-set: 9
Reviewer: Vadim Khitrin <1014336@d5d70762-12d0-45a1-890d-524b12d3f735>