Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > (1 comment)
> 
> From my personal experience, Dell's modules throw an exception when you attempt to run them remotely, setting the connection to local doesn't fix it and the task must be delegated to localhost in order to work.
> 
> The creation of in-memory inventory is done for convenience, the play aims to be user friendly and using inventory provides the best feedback for users.

Many ansible modules run in front of localhost, which is completely fine.
But I think that the creation of the inventory is useless.
It will not provide any feedback to the users because the inventory stays in memory and don't written to the file, so the user could be not aware of the inventory at all.

Patch-set: 9
Reviewer: Maxim Babushkin <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>