Update tempest config file process
Configuration of the tempest.conf should be made by providing the
required arguments during the discover-tempest-config command and not by
modifying the tempest.conf directly.

Change the subunit-trace to subunit-stats command.
The subunit-trace command is missing from RHOS 14 stestr packages.
Replace the command to subunit-stats in order to cover all the testing
versions.

Change-Id: I781e6df02113ddea65aad99c0fe184560b96bdbf