Update patch set 2
Patch Set 2: Code-Review-1

A few notes, please.
1. Change the variable to undercloud_only: true/false (false by default).
I don't see the reason to specify the full variable as it should be such by the default.

Then, in the when condition, you can specify it as a check of a boolean.

when: not undercloud_only|bool

2. Include the overcloud_nodes.yaml into the block.
It means that you will need to take the gather facts of the undercloud into the main.yml before the block.

The reason for this is that when we will create an inventory only for the undercloud we will skip the virtualenv creation and package installation which will save us a lot of time waiting.

3. Change the "Create TriplO inventory" name into some more understandable for the tasks within the file. Example - "Gather overcloud nodes details"

Patch-set: 2
Reviewer: Maxim Babushkin <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1