Update clouds.yaml path to a static location instead of current directory
When ansible-nfv is executed from undercloud, creating ~/clouds.yaml file
does not work because the ansible execution will be in ~/ansible-nfv
folder. But it will work if the ansible-nfv is executed external to the
undercloud. In order to make both cases to work (internal and external
to unercloud), use the static config path ~/.config/openstack/clouds.yaml.

Change-Id: I18c3c9e6d1194b40e638aa611837caa88b42edf9