Update clouds.yaml file generation
Some of the openstack tasks are not covered within the modules.
In order to execute them, we need to execute openstack command.
Openstack command support "--os-cloud" flag, which allow to specify the
cloud name from the file we would like to work with.

Example: "openstack --os-cloud overcloud server list"

In order to support such behavior, "identity_api_version" needs to be
added to the clouds.yaml file.

Change-Id: Iec2cd5333ef650d4c93f1ec253da3b81930fd734