Add virt_customize native Ansible modules
This patch adds WIP virt_customize_modules which are yet to be hosted on
PyPI

These modules are planned to be hosted on PyPI, until then the current
modules are not optimized and are out of scope in this review.

Tested and verified on RHEL/CentOS/Fedora.

Source of modules:
https://github.com/VKhitrin/virt-customize-ansible-modules

Change-Id: I2982fd7f98e823c67889330579f3bccb36b19d1d