Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Please do not run to Fix every ansible version changes 
> We have back port compatibility.

What do you mean?
If the new change presented, why not to update to it?
Moreover, remember that we jumped from ansible version 2.4.4 to 2.7.6.
A lot of new implementations and deprecations made in these versions.

And regarding this change, it is actually break the expected behavior.

Patch-set: 1
Reviewer: Gerrit User 1008672 <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>