Update tempest play config process
The python-tempestconf repository contains its own
discover-tempest-config tool that is available when installing the repo
with pip.
Switching to the discover-tempest-config in python-tempestconfig from
the discover-tempest-config provided within the undercloud will gives us
more flexibility and remove dependencies on the undercloud version.

The change has been tested on 13 and 14 versions.
Later, could be tested on the 10 as well.
In case works as expected, the specific tag for the osp 10 in
python-tempestconf repo could be removed.

Change-Id: Iac85640e925bc4320331a0dba65f2f15b350807c