Fix tags invocation
As of the ansible 2.5 version, the invocation of the tags has been
changed. The invocation could be done only when using the import_tasks.

As a result, in the current state the deletion of the overcloud, for
example can't be done when specifying the tags only.
The patch fix that.

Reference:
1. https://docs.ansible.com/ansible/latest/user_guide/playbooks_tags.html#tag-inheritance
2. https://docs.ansible.com/ansible/latest/porting_guides/porting_guide_2.5.html#dynamic-includes-and-attribute-inheritance

Change-Id: I65d9724f0bd98d9f54e2dd15db51108490f5033a