openstack_tasks: Disable Dynamic Inventory By Default
A known issue in 'os_floating_ip' is that a FIP is not returned as part
of a result.

The dynamic inventory creation during resource generation will fail if
no FIP was returned.

Adding a trigger which by default, will not attempt to generate a
dynamic inventory during play.

Refer to a workaround role '/roles/post_install/dynamic_host_inventory'
and to '/playbooks/packet_gen/trex/performance_scenario' for an example.

Change-Id: I4d9ea032b16edada742151e82984c3e5c6b1347c