Fix tempest playbook execution for python3
Due to Ansible version used as of now (2.7.11), Ansbile lacks the option
to detect remote python interpreter.
Python3 based deployments will fail to execute tempest playbook by
using the default '/usr/bin/python' interpreter which is not present in
RHEL8.

Set ansible_python_interpreter to '/usr/libexec/platform-python' which
is a cross distribution executable that is linked to native python
version provided in the distribution.

Change-Id: I2a2f6795f3b34854bccdabaeeda54ad132d26712