Fix selinux failure during playbook execution
The playbook runs on the localhost (control machone).
The playbook fails with the following error message:

"Aborting, target uses selinux but python bindings (libselinux-python)
aren't installed!"

The package actually installed on the localhost.
It happens when the selinux us enabled on the host.
It fails because the selinux package does not copied to the created
virtualenv and as a result, python, which runs from the virtualenv,
unable to find it.

The following fix is the most effective one and does not require
additional manual tasks that needs to be performed by the user.

More in this:
https://dmsimard.com/2016/01/08/selinux-python-virtualenv-chroot\
-and-ansible-dont-play-nice/

https://github.com/ansible/ansible/issues/34340

Change-Id: I1dc79bb9f218902005d1a8c1b50440b2dad72676