Fix selinux failure during playbook execution
The playbook runs on the localhost (control machone).
The playbook fails with the following error message:

"Aborting, target uses selinux but python bindings (libselinux-python)
aren't installed!" [1][2]

The package actually installed on the localhost.
It happens when the selinux us enabled on the host.
It fails because the selinux package does not copied to the created
virtualenv and as a result, python, which runs from the virtualenv,
unable to find it.

The solution is to install "selinux" package within the virtualenv. [3]
The package has been added to the requirements.txt file.

As a result, the "ansible_python_interpreter" is not longer required for
the tasks that are executed from the ansible control node.

More in this:
[1] - https://dmsimard.com/2016/01/08/selinux-python-virtualenv-chroot\
-and-ansible-dont-play-nice/
[2] - https://github.com/ansible/ansible/issues/34340
[3] - https://pypi.org/project/selinux/

Change-Id: I1dc79bb9f218902005d1a8c1b50440b2dad72676