Add custom_net_config.py script to handle multiple net instances
The script configures instance with one or multiple network interfaces.
It relies on the cloud-init package that perform the proper network
configuration on the first boot.

For the proper execution, the location of the script should be:
"/var/lib/cloud/scripts/per-boot/", which will cause it to run every boot.

*) During the first boot, the script dumps network configuration and stores it
   locally. During the next boot, the script takes the network dump and applies
   it on the system based on the changed (if changed) network interfaces.

*) The script set SSH configuration to "UseDNS=no" to prevent login delay.

*) The script updates the default route for the proper interface with the
   minimum metric value based on the provided "tag" by the user. The "tag"
   should be tied to the interface that the default route should be set with.

Change-Id: I27291c92f6a686e10d9f4aed2aeac4dee3ba7055