Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Is it an advanced mode of this?
> https://github.com/redhat-openstack/nfv-tempest-plugin/blob/master/nfv_tempest_plugin/tests/scenario/tests_scripts/custom_net_config.py
> I would advise put all scripts from tempest can be used under https://pypi.org/
> Instead of Code Duplication between the two plugins,
> 
> This way 1 code is maintained for two repos, and the benefit for two reps achieved

This script is a replacement of the network configuration script we are currently using in tempest plugin.

This script will be the ONLY script that will reside in both repositories (ansible-nfv and nfv-tempest-plugin).
The reason for this is that in tempest plugin it will cover instances network configuration for the tests and in ansible-nfv it will have much wide scenario usage (tests, performance and any other scenario that will require to boot an instance with one or multiple interfaces. Not related to tempest at all).

No other script of nfv-tempest-plugin will not reside in ansible-nfv repo. No reason to put it there.
When the resource creator will be used as the initial setup creator for the tempest resources, the scripts that will be passed to the instances will be given in in a separate configuration file in base64 format and will reside in nfv-qe repo (example: ci_network_config.yml).

With all the above, it looks like overkill to put one single script in pypi.

Patch-set: 1
Reviewer: Gerrit User 1008672 <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>