Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

Last testcase in lacp test plan contains a bonding in the guest. I we wanted to implement thist testcase, we would need os-net-config.

In my environment in which I only have one compute, it is interesting to configure routes in the guest (I configure 2 networks and attach 1 vm to each network and configure the switch to route from 1 network to the other one, but I also need to configure routes in the guest in order to send this traffic to the performance switch and not to the default route). It is not a standard environment, but being able to configure all these things provides a lot of flexibility.

Patch-set: 3
Reviewer: Gerrit User 1016373 <1016373@d5d70762-12d0-45a1-890d-524b12d3f735>