Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (1 comment)
> 
> Last testcase in lacp test plan contains a bonding in the guest. I we wanted to implement thist testcase, we would need os-net-config.
> 
> In my environment in which I only have one compute, it is interesting to configure routes in the guest (I configure 2 networks and attach 1 vm to each network and configure the switch to route from 1 network to the other one, but I also need to configure routes in the guest in order to send this traffic to the performance switch and not to the default route). It is not a standard environment, but being able to configure all these things provides a lot of flexibility.

=====

I have nothing against the os-net-config implementation and when you, Vadim, mentioned the bonding config within the instances, I think it for sure could help us in some usecases.
But it is not relates to this script.
The script works against the cloud-init package, which is y default installed in all rhel/centos cloud images.
So it works pretty independently and does not require any kind of pre configuration.
In addition, this script will work not just in the scenarios of tests but other as well (like perf).

And os-net-config package should be installed within the image to use it, which adds another level of complexety.

Patch-set: 4
Reviewer: Gerrit User 1008672 <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>