Update tripleo inventory template
Update the template, used to generate the inventory file.
In OSP 15, the undercloud is fully based python3.
As a result, the "ansible_python_interpreter" variable must be set in
inventory file in order to communicate with the undercloud node.

Set the "ansible_python_interpreter" to "/usr/libexec/platform-python"
if the host from the RHEL family and has a OS version 8 and above.
In case the "ansible_python_interpreter" explicitly defined, use it.

Change-Id: If3fe3214cb4bb4dc03de4cd5f1a832cc7e8472d2