Update patch set 2
Patch Set 2:

(1 comment)

It would be wonderful if you also give your view on general code. 
Instead of vlan in vif profile, should we opt for a flag to indicate we want to use vlan, since even though sriov vif has vlan in its vif profile, nic should be able to generate tagged or untagged vlan traffic.
The vlan from sriov vif profile is used as a hint for trex to traffic on vlan, the testpmd vm is on.
so should a different variable be used as indicator for vlans used by testpmd VM.

Patch-set: 2