Update patch set 1
Patch Set 1: Code-Review+2

How are we going to manage switch configuration files?
What I understand is that we should have a configuration file for CI setup together with the templates so that if we install lacp templates, this playbook will configure lacp in the switch and if we install other templates with active-backup, it will configure the ports in the normal way.
I think there should be a configuration file per setup, not a single configuration file

what do you think?

Patch-set: 1
Reviewer: Gerrit User 1016373 <1016373@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2