Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> How are we going to manage switch configuration files?
> What I understand is that we should have a configuration file for CI setup together with the templates so that if we install lacp templates, this playbook will configure lacp in the switch and if we install other templates with active-backup, it will configure the ports in the normal way.
> I think there should be a configuration file per setup, not a single configuration file
> 
> what do you think?

I see it in the following scenario.
We will have a separate branch in the nfv-qe repo.
The branch will hold the configuration file that configures the switch in the standard manner.

In case of a required change, a person will fire a patch to the gerrit system.
Once approved, gerrit will trigger a playbook execution that will run and configure the switch in the new configuration.

Regarding lacp test. Currently, the test requires a separate job, but in the future, once all tests will switch to use the resource creator, the test will be a part of the regular deployment.
It means that the switch configuration for lacp should be change before the test and after the test, be reverted back to the standard.

Patch-set: 1
Reviewer: Gerrit User 1008672 <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>