Update patch set 3
Patch Set 3: -Code-Review

> Patch Set 3:
> 
> @Miguel, please explain how the following patch are related to your introspection phase issues?

It is the same issue, but in the hypervisor. It is installed version 20 that it is not working as it is installed python 2.7.
But the think is that I am not sure who is installing it. Is it inside ansible-nfv?

Patch-set: 3
Reviewer: Gerrit User 1016373 <1016373@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0