Update patch set 1
Patch Set 1:

Does it required template changes for configs? Since we are changing the type of an existing variable, it would be good to have fail message, when properities is list (for users, who are using it explicity).

Patch-set: 1
Reviewer: Gerrit User 1010167 <1010167@d5d70762-12d0-45a1-890d-524b12d3f735>