Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> The file should not be added in the root directory itself. if you could provide the description or usage of the file, then we could decide on a better location of the file.

Sure, I ll attach a README for it, but before that, I need to know the suitable folder in ansible-nfv to move this script, as it is the first ever requirement to run perf playbook in different configurations of VM and DuT, and that is the purpose of having this tool to automate configurations, execution and results collection. I suggest a folder script/ to group nfv_perf_mq and its README. Any other suggestions ?.

Either before merge (if I get access to CI lab) or after merge (assuming it wont be enabled until following is done), I wished it to keep it ready for target CI.

We need to change only the test globals section in it, that I can update in v2 of this patch without testing it or test if when I have access (but I do not know when I get one) .

Thanks.

Patch-set: 1