Update patch set 1
Patch Set 1:

Looking at the script, the first impression is that it fits the description of an ansible rol. Any specific reason to keep it as bash instead of creating a ansible role?

Also I would like to know how and when this will be invoked.

Patch-set: 1
Reviewer: Gerrit User 1010167 <1010167@d5d70762-12d0-45a1-890d-524b12d3f735>