Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Looking at the script, the first impression is that it fits the description of an ansible rol. Any specific reason to keep it as bash instead of creating a ansible role?
> 
> Also I would like to know how and when this will be invoked.

No reason to start with something easier in bash! than writing playbook! .. We can write a playbook, but this script would then be replaced by that time it is available. Ack ?.

Up to QE to decide on triggering it, but I hope it will be triggered following perf run for single queue ovs-dpdk perf CI.

Patch-set: 1