Update patch set 2
Patch Set 2: Code-Review-1

I totally agree with Saravanan.
I think that the script should be converted to a role or even to a number of roles that could inherit one from another.

Most of the commands in the script are executed by using the ansible ad hoc command by using the shell module.
Many of the commands have the support of the native modules.
Why not to use them?
The advantage to use the native modules (when available) is a proper error handling and readability.

Moreover, I think that when converted to a role it will be much easier to read and understand the flow of the tasks.

In my opinion, the location of the role should be in "roles/packet_gen", as related directly to the perf scenarios.

Patch-set: 2
Reviewer: Gerrit User 1008672 <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1