Add a base structure to create provision for adding tests
molecule can be used to test the ansible roles standalone.
It can be used with container (docker or podman) or vm (vagrant).
This patch adds a skeleton to create role with molecule structure
and a vagrant based test scenario for tuned coniguration. This is
only a demo, the test willbe completed as a follow-up.

Change-Id: Ieb885a91352392e35183b6dbff92ce8d0b13f146