Add molecule testing to openstack_tasks role
Molecule [1] is a testing framework adopted by Ansible.
As a first integration of molecule in ansible-nfv repo, the first role
to adapt the testing framework is - openstack_tasks role.

As for now, ansible-nfv will use molecule version - 2.22.
It is not the latest version and the reason for this is that the next
version of molecule is the major version change and it makes the
separation of the drivers from the core.
The separated drivers in the new version are still at the early stages
of development and as a result, we will use the older version.

[1] - https://molecule.readthedocs.io/en/2.22/index.html

Change-Id: Icaa37fa7d24e40f29ffd7ae69e15c6c3c62de490