Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Thanks for introducing,
> I do not have clear view for the necessity, is it for gating?
> What is necessity pf the patch?

Yes, it is the first step (poc) in covering our tooling with tests.
This patch is the ansible-nfv part of testing.
The next patch will be for jenkins and create the job that will use this framework for the patch testing.

In the coming qe internal meeting I will make a more deeper introduction of molecule and answer your questions.

Patch-set: 2
Reviewer: Gerrit User 1008672 <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>