Fix performance play failure
In the performance playbook, hci testing is an optional testing flow.
In the standard perf execution, "hci_group" will not exists.
Set the group to "omit" in order to not fail the play in case the flow
not executes hci.

Change-Id: Ic48906f3eae69b0fb935e7a12b1ec85391c22970
(cherry picked from commit 79cb7b1054e3d772af566ded26dcd06f9ba42822)