Update patch set 1
Patch Set 1:

> Each user, will override the default vars and build templates for his own deployment.

Will those user's vars file will be maintained in this repo (or) a common internal repo (or) it will be left to the user? It is importatnt that it is maintained in a common place, which allows anyone to use any setup.

Patch-set: 1
Reviewer: Gerrit User 1010167 <1010167@d5d70762-12d0-45a1-890d-524b12d3f735>