Update the driver of openstack_tasks and tempest molecule roles
Update the driver of "openstack_tasks" and "tempest" molecule roles from
"openstack" driver to "delegated" driver.

The driver in molecule used to specify which resource is used to create
test instances. Since we are not using the actual api of openstack
environment to create test instnaces, "openstack" driver is not needed
here.
The "openstack_tasks" and "tempest" roles does not require any test
instance to be created so the "delegated" driver is used.
The "delegated" driver moves the responsability of the creation of the
test environment and inventory to the user.

Change-Id: I46b1706c7a4fd3225b63449f76d8a43376c35295