Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> On a general feedback, I'm still not yet convinced by adding yet another layer of indirection, especially when we want to add/modify a parameter.
> We already have a lot to know/understand with TripleO, and we now have another parameter mapping to have in mind.
> Can't we use directly the same name as in TripleO ?
> For instance, to use "NeutronMechanismDrivers" instead of "neutron_mechanism_drivers".
> 
> Also, it seems we exchange 1 big file to another, bigger file. Can't we split this to 1 feature/file ?
> 
> On the nic-config file, how will you handle DCN ?

The whole point of this initiative as a general is to reduce the amount of the templates we need to maintain. For me, it is really annoying to change one single parameter in 7-8 duplicate template scenarios.

Regarding the same names of the parameters as in TripleO, sounds reasonable to me. Will do it.

The features are splitted into separate file. DPDK and SRIOV resides both in separate files.

What is about DCN? How is it different?

Patch-set: 2
Reviewer: Gerrit User 1008672 <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>