Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > The whole point of this initiative as a general is to reduce the amount of the templates we need to maintain. For me, it is really annoying to change one single parameter in 7-8 duplicate template scenarios.
> > 
> > The features are splitted into separate file. DPDK and SRIOV resides both in separate files.
> > 
> 
> That's the point of using 1 file per feature that you *reuse* across multiple template scenarios.
> This way, you modify the parameter in 1 place, and the modification will be used across the scenarios.
> 
> Here, we *generate* multiple files for multiple scenarios: the split of 1 feature / file doesn't bring any real value as we won't read the generated files.

If the file will not needed, it will not be generated.
The generation of the files done based on the users input and not as a full bulk.

> 
> > What is about DCN? How is it different?
> Have a look at the existing templates, we have different a copy of each network per leaf for OSP13.
> And this has evolved with OSP16.

I don't see any problem with the dcn templates.
I logic of nic-configs generation based on the type of the resource we would like to add (interface/linux_bond/ovs_bridge/ovs_user_bridge). If any additional type will be needed, we could easily add it.

Patch-set: 2
Reviewer: Gerrit User 1008672 <1008672@d5d70762-12d0-45a1-890d-524b12d3f735>