Tripleo inventory - Remove old env directory
When generating the inventory on a setup and then later regenerate the
inventory on a re deployed setup, the playbook flow may fail because it
will conflict with the existing environment directory.

In order to prevent such conflict, remove the host environment directory
at the beginning on each inventory generation.

Change-Id: I38068220d47f481bec6c6afcbadea53b6bc1f6a6