flatten filter disorder port names
Depending on the port names, flatten could disorder port names.
Later on, with this port names it is taken dut_macs that is used as
a parameter in binary_search command. If the order of the macs is
wrong, no traffic will arrive to testpmd.

In our templates, all of our port names are alphabetically ordered,
so in order not to have them disordered by the flatten filter,
we can ensure that they are ordered using the sort filter.

This is a example of how flatten disorder:

Input: management_net_137, mlnx_net_nic0_128, mlnx_net_nic1_129

Output: mlnx_net_nic1_129, management_net_137, mlnx_net_nic0_128

Output with sort: management_net_137, mlnx_net_nic0_128, mlnx_net_nic1_129

Change-Id: Id3e04051b6ff39b28bf18f0a8b495bc05c962e3f