Tempest - Add regex filtering to the stestr tests execution
Modify tempest role to pass tests regexes or external file instead of
looping through the list of tests.
Add the following variables:
- tests_regex: Global tests regex filter
- tests_include_list: Path to the list with regexes or tests
New parameters are used for the stestr execution.

Remove "tempest_tests" variable that replaces with the above.

Using a regex reduces the maintenance of the tests inputs as per now it
will be provided as a regex and not per each test definition, so the new
tests uder the existing scenario will be added automatically, unless
specified in the regex to be skipped.

Change-Id: I3ecfd3ab749dcf3d435a47cfe0f53b8b3edd8896