post_install: Refactor volume creation
Volume creation used to be only available when working with dynamic instance
creation role. Refactoring this to expose volume creation as a generic resource
creator option.

This change breaks prepare_dynamic_instances role which was implemented as
part of WIP of multi DuT performance which is not implemented.
Removing this role from performance automation until it will be refactored.

Change-Id: I0d0a4fae05a38795cb4f6d7dcf243e976e183aa2