Fix pip packages install issue
The openstacksdk pip package fails to be installed because it tried to
be installed before pip and setuptools packages updated.
When the packages in a pip install tasks defined in a list, it does not
mean they will be installed in that order.

Separating the update of pip and setuptools from the other packages
install.

Change-Id: Ifabd87d704d83e3de26875e25e50b66a3c0eb722