Fix molecule gate random failure
The gate fails randomly on verification of the networking section for
the openstack_tasks role. I faced the same issue while running the gate
test manually.

The reason for the random failure, is that during the molecule
verification, at networking section, existing networks fetched and
exising settings compared to the provided configuration file.

During the last patch of security groups, second network was added to be
able to test "security_groups" flag during network creation.

While network details gathered two internal network randomly swapped
between them and as a result, configuration was not reflecting the
testing configuration.

In order to not overcomplicate the molecule configuration for the
openstack_tasks role:
- Removing the second network.
- Adding the "port_security: true" flag to test the security group
  network implementation.
  Changing the flag from "false" to "true" will not make any difference
  because the module will check provision of the flag from the config
  file and not use api default.

Change-Id: I786228941798deb1acb5cc38467eba5f0f9572ba