Tripleo_inventory - Add ssh arguments for connection
When tripleo_inventory role creates an inventory file for the hybrid
environment, it adds ssh arguments to avoid host checking.
But it adds it only for the proxy server (hypervisor) and not the
overcloud ndoes or undercloud.

Resently "jump" tool started to fail during execution.
The reason for the failure - jump unable to identify the puddle id of
the deployment.
Identification of the puddle id, performed by jump tool by running the
ad-hoc command using the inventory file and "cat" the content of the
core_puddle_version file.
The error was the following:
"host key for 172.16.0.47 has changed and you have requested strict checking"
The ip address is the address of the undercloud vm.

After debugging, it seems that this issue happens only when
tripleo_inventory is used to generate the inventory for the job.
When the deployment performed from scratch and uses infrared to generate
the inventory file, the issue does not seems to be happen.
(At least from by search on the failed jobs in jenkins).

When I compaired the inventory files by tripleo_inventory and by
infrared, I noted that we are using "StrictHostKeyChecking=no" and
"UserKnownHostsFile=/dev/null" ssh arguments only for proxy command,
which is applied to the jump host (hypervisor) but not the overcloud
nodes, while infrared inventory have these options for both.

That's pretty weird because the tempest and perf playbooks does not
having any issues on connecting to the undercloud host, but maybe there
is some difference between using ansible-playbook and ansible as ad-hoc
command.

That is the only possible reason I found till now.

Now, to the changes:
- Add "StrictHostKeyChecking=no" and "UserKnownHostsFile=/dev/null" to
  eliminate possible connection fingerprint issues.
- Add "ForwardAgent=yes" to use the ssh key forwarding.
- Add "ServerAliveInterval=30" to make the client wait before sending
  a null packet to the server (to keep the connection alive).
- Add "ControlMaster=auto" to enable the sharing of multiple sessions
  over a single network connection.
- Add "ControlPersist=30" connection will remain open for the
  specified number of seconds.
- Add "BatchMode=yes" to fail immediately if the key is rejected,
  instead of failing back to a password prompt.

Change-Id: I815bfec2b36af6141abd7ee1fe19cb7c0e40306f