Tripleo-inventory - Add infrared plugin definition
Currenty infrared is using tripleo-inventory plugin from a different
source than ansible-nfv repository.
That requires higher maintenance and duplication of the role.

Adding infrared plugin definition to provide the ability to use the role
as infrared plugin.
After the the patch will be merged, infrared plugin install dict will be
updated to install the plugin automatically from ansbile-nfv and not
perform it manually.

Change-Id: I3a73ae23e4a9ea53b4c801f3e19157f89b930c50