Openstack_tasks - Add private key path as variable
During creation of the resources with openstack_tasks role, private key
that is used to connect to the instances, fetched to the local machine
under the "/tmp" directory.
When parameterized jenkins jobs flow was implemented, same jenkins slave
was able to execute different jobs. As a result, if both jobs are
executing performance playbook with resources creation, private key
fetched to a jenkins slave may got overridden because fetched into the
same "/tmp" directory with possible same name.

Creating a variable called "private_key_fetch_location" that will allow
to set a custom path for the private key to be fetched.

Change-Id: Idcd4805e7f56ef61379d61ac570ebc19685c34a2