Add support for TRex only deployment (shift-on-stack)
Add support to configure dut_group and dut_compute
to be empty values, so that TRex server can be deployed
without testpmd vm. Allows to send traffic to testpmd,
which is running inside OCP on OSP.

Add a check to skip floating ip configuration, if the
any one of the instaces already has floating ip. In the
current implementation of os_floating_ip module, it does
not validate if it already exists. Added a safe condition
to allow re-runs. Note that re-runs are used to create
trex vms so that port's mac address can be provided to the
testpmd inside OCP. In the 2nd run trex to start packet
generation with destination mac address.

Add bool filter to the conditions checks to allow it to run
by disabling trex run. In the current implementation, even it
is set as false, since the bool filter is not used, it is treated
as valid string.

Change-Id: I7757572f660d4fdbbf914709fe10bb87a5704f4e