Avoid permission change to existing directories
When using openstack_tasks role, clouds.yaml file created in order to
manage and control the relevant stack/project/user resources.

By default, the clouds.yaml file created in the
~/.config/openstack/clouds.yaml path.
But in a freshed deployed setup, the "openstack" directory does not
exists. It means that the directory needs to be created prior the
clouds.yaml file creation.

In the shiftstack scenario, the location of the clouds.yaml file should
be within the "/home/stack" directory. But because the directory where
the clouds.yaml file located are pre-created with the "755" permissions,
it leads to the state where stack home directory changes the permissions
from "700" to "755" which leads to the ssh connection failed because of
the wrong home directory user permissions.

Adding the check for the directory existence.
If directory exists, the directory will not be created and permissions
will not be changed.
If directory does not exist, directory will be created and permissions
will be set.

Change-Id: Ibef9c2217ef246ab032b28bcc817d1f1316e9ce0