Refactor tempest role
Tempest role consists of the main file that contains 472 lines.
It's a huge file that is hard to read and understand.
The file contains execution for both git tempest and containerized
tempest.

Split the main file into smaller parts:
- prep_container_tempest
- prep_git_tempest
- config_exec_container_tempest
- config_exec_git_tempest

Place all tempest related config files in the single "tempest" dir.

Benefits of the change:
- Easy to read and understand the role
- When the playbook executes, because of the "include_tasks" way only
  the executed tasks are shown and it not shows a lot of "skipped"
  tasks.
- Files of the tempest will be placed in one single directory which will
  make in logically structured.

Change-Id: Ia0fa4d4c441eb52e199121cf3cbafd68680fd214