Add option --disable-upward-search to binary search
If the final rate is the same than the specified in binary search
command, I get the message:
Detected requested rate is too low, doubling rate and restarting search

And start searching for higher values. If as in our case, trex is limiting the
rate, it will waste a long time looking for other rates and at the end it
gives a very low rate, i think it is caused because it injects very high rates
and overflow buffers.
With this option enable, it will not try higher rate values than the specified

Change-Id: Ia2b6b91daadce9127bc9a61b811336b74e07b14c