cpu pinning not configured in performance vms
Modified ansible tasks. I have created  2 steps:
1. remove old configuration
2. create new configuration

The reason to use 2 steps is that with a single step, if the
parameter didnt exist before, it will not match the regex and
the parameter will not be added. Only already existing parameters
will be updated.

With the 2 steps, the parameter will be added to the file always,
even if it didnt exist before

Change-Id: I4c5b57ece44ad051e796e1e800ea5228e4014daf