[ wip] multique playbook
It will generate a json file with the flows that each queue will receive
This file could be used by tempest testcase to send traffic to
specific queues.
File will be located in tres file and will be the input of traffic
generator multiqueue.py. File has the following structure:
[
    {
        "name": "eth1",
        "queues": {
            "2": {
                "pps": 0.2,
                "isg": 0,
                "ips": [
                    "48.0.0.26",
                ]
            },
            "1": {
                "pps": 0.2,
                "isg": 0,
                "ips": [
                    "48.0.0.1",
....
It contain a list of ips for each port and queue

Change-Id: Ib694e3ab0c13e5ea0d34415a55e179ea25308da7