ansible-lint: Bump Version To 5.3.2
Due to https://github.com/ansible-community/ansible-lint/issues/1795
we are required to move to a newer ansible-lint version.

Version 5.0.0 refactored the rules and we must adapt to the new rules.

Updated .ansible-lint.
Excluding the playbook 'playbooks/packet_gen/iperf/performance_scenario.yml'
due to a linting fail that we can not satisfy.

Excluding rules:
  - 'unnamed-task' # We don't mind having tasks that are unnamed
  - 'no-changed-when' # We don't mind that there will be tasks that will always return ok
  - 'no-handler' # We don't mind that we're running tasks based if a previous task returned changed
  - 'package-latest' # We don't mind installing latest packages\

Change-Id: I057faab90874618933d935bb77ff28663f22e3f0